+++ template/lugia/class/view/modals/DetailModal.js
@@ -117,7 +117,7 @@ class DetailModal extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
+					getPopupContainer: () => document.getElementById('$pageClassNameDetailModal'),
 					allowClear: true,
 				},
 			},
@@ -141,7 +141,7 @@ class DetailModal extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameDetailModal'),=
+					getPopupContainer: () => document.getElementById('$pageClassNameDetailModal'),=
 				},
 			},
 			{
@@ -158,7 +158,7 @@ class DetailModal extends PureComponent {
 					disabled: showdetails,
 					placeholder: '请选择日期',
 					initialValue: moment(),
-					getCalendarContainer: () => document.getElementById('$PageNameDetailModal'),
+					getCalendarContainer: () => document.getElementById('$pageClassNameDetailModal'),
 				},
 			},
 			{
@@ -201,7 +201,7 @@ class DetailModal extends PureComponent {
 			}
 		];
 		return (
-			<div id='$PageNameDetailModal'>
+			<div id='$pageClassNameDetailModal'>
 				<NeatForm
 					refs={(ref) => (this.createProduct = ref)}
 					labelSize='100px'