+++ template/lugia/hooks/models/dataModel.js
@@ -6,21 +6,19 @@
 import { page } from 'yss-trade-base';
 
 export default {
-	TableList: [], //列表
-	TableListTotal: 0,
-	//获取列表
+	TableList: [], //主列表
+	TableListTotal: 0, //主列表数据总数
+	//获取主列表入参
 	queryTableList: {
 		...page,
-		startDate: null,
-		endDate: null,
 		reqPageSize: 50,
 	},
 	/***弹框status*/
 	isOpenFormModal: {
-		page: '$PageName',
-		type: 'add', //add, update, delete, import
+		page: '$pageClassName',
+		type: 'add', //add, update, delete, import, log
 		status: false,
 	},
-	/**保存选择行信息*/
+	/**保存主列表当前选择行信息*/
 	projectRowed: {},
 };