+++ template/lugia/hooks/view/components/MainTable.js
@@ -60,7 +60,7 @@ const MainTable = (props) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameMainTable'),
+				getPopupContainer: () => document.getElementById('$pageClassNameMainTable'),
 			},
 		},
 		{
@@ -71,7 +71,7 @@ const MainTable = (props) => {
 			type: 'RangePicker',
 			props: {
 				allowClear: false,
-				getCalendarContainer: () => document.getElementById('$PageNameMainTable'),
+				getCalendarContainer: () => document.getElementById('$pageClassNameMainTable'),
 			},
 		},
 		{
@@ -92,7 +92,7 @@ const MainTable = (props) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameMainTable'),
+				getPopupContainer: () => document.getElementById('$pageClassNameMainTable'),
 			},
 		},
 	];
@@ -108,7 +108,7 @@ const MainTable = (props) => {
 		e.stopPropagation();
 		changeSync({
 			isOpenFormModal: {
-				page: '$PageName',
+				page: '$pageClassName',
 				type: 'add',
 				status: true,
 			},
@@ -189,7 +189,7 @@ const MainTable = (props) => {
 		e.stopPropagation();
 		changeSync({
 			isOpenFormModal: {
-				page: '$PageName',
+				page: '$pageClassName',
 				type: 'detail',
 				status: true,
 			},
@@ -202,7 +202,7 @@ const MainTable = (props) => {
 		e.stopPropagation();
 		changeSync({
 			isOpenFormModal: {
-				page: '$PageName',
+				page: '$pageClassName',
 				type: 'update',
 				status: true,
 			},
@@ -372,7 +372,7 @@ const MainTable = (props) => {
 	return (
 		<Fragment>
 			{/* 查询表单 */}
-			<div id='$PageNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
+			<div id='$pageClassNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
 				<SearchForm
 					refs={(ref) => {
 						searchForm.current = ref;
@@ -425,7 +425,7 @@ const MainTable = (props) => {
 					}}
 				/>
 				{withRoleBotton(ButtonType)}
-				<ReactTable {...getTableConfig()} resizeTableCode='MainTable-$PageName' tableCode='pingan-MainTable-$PageName' />
+				<ReactTable {...getTableConfig()} resizeTableCode='MainTable-$pageClassName' tableCode='pingan-MainTable-$pageClassName' />
 			</div>
 			{/***弹框组件** */}
 			<Modal
@@ -441,7 +441,7 @@ const MainTable = (props) => {
 				onCancel={() => {
 					changeSync({
 						isOpenFormModal: {
-							page: '$PageName',
+							page: '$pageClassName',
 							type: 'add',
 							status: false,
 						},