+++ template/lugia/hooks/view/modals/DetailModal.js
@@ -121,7 +121,7 @@ const DetailModal = (props, ref) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
+				getPopupContainer: () => document.getElementById('$pageClassNameDetailModal'),
 			},
 		},
 		{
@@ -144,7 +144,7 @@ const DetailModal = (props, ref) => {
 					maxHeight: '400px',
 				},
 				dropdownMatchSelectWidth: false,
-				getPopupContainer: () => document.getElementById('$PageNameDetailModal'),
+				getPopupContainer: () => document.getElementById('$pageClassNameDetailModal'),
 			},
 		},
 		{
@@ -161,7 +161,7 @@ const DetailModal = (props, ref) => {
 				disabled: showdetails,
 				placeholder: '请选择日期',
 				initialValue: moment(),
-				getCalendarContainer: () => document.getElementById('$PageNameDetailModal'),
+				getCalendarContainer: () => document.getElementById('$pageClassNameDetailModal'),
 			},
 		},
 		{
@@ -205,7 +205,7 @@ const DetailModal = (props, ref) => {
 	];
 
 	return (
-		<div id='$PageNameDetailModal'>
+		<div id='$pageClassNameDetailModal'>
 			<NeatForm
 				refs={(ref) => (createProduct = ref)}
 				labelSize='100px'