+++ template/redux/class/view/components/MainTable.js
@@ -57,7 +57,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
+					getPopupContainer: () => document.getElementById('$pageClassNameMainTable'),
 				},
 			},
 			{
@@ -68,7 +68,7 @@ class MainTable extends PureComponent {
 				type: 'RangePicker',
 				props: {
 					allowClear: false,
-					getCalendarContainer: () => document.getElementById('$PageNameMainTable'),
+					getCalendarContainer: () => document.getElementById('$pageClassNameMainTable'),
 				},
 			},
 			{
@@ -89,7 +89,7 @@ class MainTable extends PureComponent {
 						maxHeight: '400px',
 					},
 					dropdownMatchSelectWidth: false,
-					getPopupContainer: () => document.getElementById('$PageNameMainTable'),
+					getPopupContainer: () => document.getElementById('$pageClassNameMainTable'),
 				},
 			},
 		];
@@ -110,7 +110,7 @@ class MainTable extends PureComponent {
 			e.stopPropagation();
 			dispatchUpdateStore({
 				isOpenFormModal: {
-					page: '$PageName',
+					page: '$pageClassName',
 					type: 'add',
 					status: true,
 				},
@@ -182,7 +182,7 @@ class MainTable extends PureComponent {
 			e.stopPropagation();
 			dispatchUpdateStore({
 				isOpenFormModal: {
-					page: '$PageName',
+					page: '$pageClassName',
 					type: 'detail',
 					status: true,
 				},
@@ -195,7 +195,7 @@ class MainTable extends PureComponent {
 			e.stopPropagation();
 			dispatchUpdateStore({
 				isOpenFormModal: {
-					page: '$PageName',
+					page: '$pageClassName',
 					type: 'update',
 					status: true,
 				},
@@ -248,7 +248,7 @@ class MainTable extends PureComponent {
 						e.stopPropagation();
 						dispatchUpdateStore({
 							isOpenFormModal: {
-								page: '$PageName',
+								page: '$pageClassName',
 								type: 'log',
 								status: true,
 							},
@@ -387,7 +387,7 @@ class MainTable extends PureComponent {
 		return (
 			<Fragment>
 				{/* 查询表单 */}
-				<div id='$PageNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
+				<div id='$pageClassNameMainTable' style={{ padding: '0px 20px', position: 'relative' }}>
 					<SearchForm
 						formItem={formItems}
 						labelSize={'70px'}
@@ -444,9 +444,9 @@ class MainTable extends PureComponent {
 					<ReactTable
 						ref={(ref) => ($mainTable = ref)}
 						{...getTableConfig()}
-						resizeTableCode='MainTable-$PageName'
-						tableCode='pingan-MainTable-$PageName'
-						className='pingan-MainTable-$PageName'
+						resizeTableCode='MainTable-$pageClassName'
+						tableCode='pingan-MainTable-$pageClassName'
+						className='pingan-MainTable-$pageClassName'
 					/>
 				</div>
 				{/***弹框组件** */}
@@ -459,7 +459,7 @@ class MainTable extends PureComponent {
 					onCancel={() => {
 						dispatchUpdateStore({
 							isOpenFormModal: {
-								page: '$PageName',
+								page: '$pageClassName',
 								type: 'add',
 								status: false,
 							},